Moved the DragPreviewContext into the Column component which is the one setting the value of the context
