--- src/management-system-v2/app/(dashboard)/[environmentId]/processes/[processId]/_user-task-builder/_utils/drag-preview-context.tsx
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const DragPreviewContext = createContext(false);
-
-export default DragPreviewContext;
